srvs/extractor/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==2.2.2
Werkzeug==2.2.2
grpcio==1.62.0
protobuf==4.25.3
protobuf3==0.2.1
numpy==1.24.3
opencv-python==4.9.0.80
paramiko==3.4.0

